Added Correlation alignment
CORAL models added. The only difference here is the output layer
in the generator is 512, and so the input is 512. Since we calculate
the covariance, we can't do 508*32 like the other models.

Completed train and evaluation functions and run too.

I think the implementation is either wrong or the model doesn't work
very well. Guess which one I believe... So I'll have to validate and fix
it later. Or we add another method for UDA instead of deep coral.